rename: EditTrackingConfigModal, EditSyncConfigModal
